Add a non-generic Vector implementation to test specialization.
The finding is as expected: specialization is working pretty well everywhere
except for map, where the combination of class and method specialization
totally fails to do anything useful.

This can easily be observed in the bytecode: for Vector[Int].map[Int] we call
Function1#apply, whereas we'd like to be calling Function1#apply$mcII$sp
instead.