Fundamentally restructured debox.map.Map[A, B].
This change fixes a longstanding (and fundemantal) bug that existed in
remove(). It also introduces some nice inlining macros, and makes use
of @tailrec to try to optimize many of the loops.

There is still probably some work to be done tightning this code up
(and measuring trade-offs) but for now it seems to work.