Push Map refactor through to Set, with extension methods.
This commit begins to explore the design talked about on the Scala
mailing lists: having a relatively small core of functionality with
extension methods (supplied by implicits) and macros to add
additional functionality in a way that minimizes megamorphic virtual
methods.

This is still a work in progress but there are already some promising
results... see the Map and Set benchmarks for examples of the macros
at work.