Refactor Buffer, fix some more bugs, add more tests.
Now we're starting to get somewhere! This commit removes the
idiosyncratic Mutable/Immutable classes and replaces them
with a single, mutable Buffer. Seriously, if you want immutable
types, Debox is not (yet) the project for you. Maybe we'll
eventually get there, but for now we need to take baby steps.